Removed HugsLib dependency
- Removed HugsLib dependency
