+++ Source/better jump pack/BetterJumpPackMod.cs
@@ -1,0 +1,16 @@
+﻿using HarmonyLib;
+using System.Reflection;
+using Verse;
+
+namespace betterJumpPack
+{
+    public class BetterJumpPackMod : Mod
+    {
+        
+        public BetterJumpPackMod(ModContentPack content) : base(content)
+        {
+            new Harmony("com.yayo.BetterJumpPack").PatchAll(Assembly.GetExecutingAssembly());
+        }
+        
+    }
+}
